Split out services into module
