+++ libriscan/biblios/services/suggestions.py
@@ -1,0 +1,4 @@
+from spellchecker import SpellChecker
+
+spell = SpellChecker()
+
